Move files a bit and remove `control` word from filenames
