+++ packages/ui-informed/src/controls/checkbox/Checkbox.stories.tsx
@@ -1,21 +1,20 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { InputControl } from './InputControl';
-import { Form } from '..';
+import { Checkbox } from './Checkbox';
+import { Form } from '../..';
 
-const meta: Meta<typeof InputControl> = {
-  component: InputControl,
+const meta: Meta<typeof Checkbox> = {
+  component: Checkbox,
   tags: ['autodocs'],
-  title: 'Controls/InputControl',
+  title: 'Controls/CheckboxControl',
   args: {
-    name: 'value',
-    label: 'Input',
-    placeholder: 'Placeholder',
+    name: 'input',
+    label: 'Checkbox',
     tooltip: 'Tooltip',
   },
   decorators: [
     Story => (
-      <Form initialValues={{ value: '' }}>
+      <Form initialValues={{ input: false }}>
         <div className='flex flex-col sm:col-span-3'>
           <Story />
         </div>
@@ -26,7 +25,7 @@ const meta: Meta<typeof InputControl> = {
 
 export default meta;
 
-type Story = StoryObj<typeof InputControl>;
+type Story = StoryObj<typeof Checkbox>;
 
 export const Basic: Story = {
   args: {},
