+++ packages/ui-informed/src/controls/checkbox/Checkbox.tsx
@@ -1,21 +1,21 @@
-import { Checkbox } from '@utima/ui';
+import { Checkbox as CheckboxUI } from '@utima/ui';
 import type { ComponentProps } from 'react';
 
-import { FormControl } from './formControl/FormControl';
+import { FormControl } from '../../formControl/FormControl';
 
-type CheckboxControlProps = Omit<ComponentProps<typeof FormControl>, 'render'>;
+type CheckboxProps = Omit<ComponentProps<typeof FormControl>, 'render'>;
 
 /**
  * CheckboxControl component that is controlled by Informed. It is a wrapped in the
  * `FormControl` component, which provides the necessary props for Informed to
  * work along with label and error message handling.
  */
-export function CheckboxControl({ ...restProps }: CheckboxControlProps) {
+export function Checkbox({ ...restProps }: CheckboxProps) {
   return (
     <FormControl
       {...restProps}
       render={({ field: { ref, userProps, fieldApi, fieldState } }) => (
-        <Checkbox
+        <CheckboxUI
           ref={ref}
           value={fieldState.value as any} // TODO: correct type
           onCheckedChange={value => {
