+++ packages/ui-informed/src/controls/input/Input.tsx
@@ -1,31 +1,25 @@
-import { TextArea } from '@utima/ui';
+import { Input as InputUI } from '@utima/ui';
 import type { ComponentProps } from 'react';
 
-import { FormControl } from './formControl/FormControl';
+import { FormControl } from '../../formControl/FormControl';
 
-export type TextAreaControlProps = Omit<
-  ComponentProps<typeof FormControl>,
-  'type' | 'render'
-> & { rows: number };
+type InputControlProps = Omit<ComponentProps<typeof FormControl>, 'render'>;
 
 /**
- * TextArea component that is controlled by Informed. It is a wrapped in the
+ * Input component that is controlled by Informed. It is a wrapped in the
  * `FormControl` component, which provides the necessary props for Informed to
  * work along with label and error message handling.
  */
-export function TextAreaControl({
-  rows = 4,
-  ...restProps
-}: TextAreaControlProps) {
+export function Input({ type = 'text', ...restProps }: InputControlProps) {
   return (
     <FormControl
-      type='text'
       {...restProps}
+      type={type}
       render={({ field: { ref, userProps, informed, fieldState } }) => (
-        <TextArea
+        <InputUI
           ref={ref}
-          rows={rows}
           variant={fieldState.showError ? 'danger' : 'default'}
+          type={type}
           {...userProps}
           {...informed}
         />
