+++ packages/ui-informed/src/form/index.ts
@@ -3,8 +3,8 @@
  * to allow for dot notation imports.
  */
 export { Form as Root } from './Form';
-export { CheckboxControl as Checkbox } from '../controls/CheckboxControl';
-export { InputControl as Input } from '../controls/InputControl';
-export { SelectControl as Select } from '../controls/SelectControl';
-export { SelectItem as SelectItem } from '@utima/ui';
-export { TextAreaControl as TextArea } from '../controls/TextAreaControl';
+export { Checkbox } from '../controls/checkbox/Checkbox';
+export { Input } from '../controls/input/Input';
+export { Select } from '../controls/select/Select';
+export { SelectItem } from '@utima/ui';
+export { TextArea } from '../controls/textArea/TextArea';
